requirements.txt
flask
flask-socketio
eventlet
yfinance
tensorflow
pandas
numpy
python-dotenv
yagmail
